Implemented the endless currency number system and menu
Endless currency is stored as strings and a package was created with addition, subtraction, formatting and comparing functions for these large numerical strings. The menu has buttons in place and a framework for opening further panels that will contain interactible, specific actions e.g. turn sound off in the settings panel